Update ID translation to latest :content:
Imagine a human website. 

Imagine a place where dank memes flow like honey.

Imagine a place where potatoes thrive. 

Imagine a place where humans and bots live in harmony. 

Imagine, in other words, a place like Discord. 

As part of its latest ad campaign, Discord asked users to describe the chat platform to someone who had never used it before. It’s a testament to the company’s customer love and broad usage that the results read like dispatches from an unknown and invisible realm; Lucy and Edmund trying to explain Narnia, Harry relaying the magic of the wizarding world, Neo stumbling through a depiction of the The Matrix. 

Imagine a closet that opens onto a kingdom of talking animals. 

Imagine running through a wall to find your new school. 

Imagine a pill that melts reality. 

The fact that these work just as well for Discord as the official entries tells us something subtly profound. Discord is not just a communication platform, it’s a hidden world.