Update ID translation to a sigma update
Here on Discord, for the first time ever I was able to find multiple people that played many nostalgic games that I played as a kid. When you’re very young, memories seem to be… I’m not sure how to word it. Childhood games that I played really aren’t that good, but for some reason I just absolutely love the memories and looking back at them, no matter how good or bad they were. On Discord for the first time I got to find others that played some of the same games I did and it was such a throwback talking about it.